Fixed Population Value in Parser
Population values are now reading in correctly from the parser, and cutoff population sizes were calculated/applied for the city match calculation.

+++ resources/src/City.h
@@ -8,12 +8,12 @@ class City{
 private:
     string city_name;
     string country;
-    double population;
+    int population;
     double latitude;
     float city_match = 0.0;
 
 public:
-    City(string city_name, string country, double population, double latitude);
+    City(string city_name, string country, int population, double latitude);
     void match_city(string& country_pref, int pop_pref, int lat_pref);
     float get_city_match();
     double get_pop();
